+++ src/wallet/wallets/metamask-injective/types.ts
@@ -1,0 +1,7 @@
+import type { MetaMaskInpageProvider } from "@metamask/providers";
+
+export type Ethereum = MetaMaskInpageProvider;
+
+export type Window = {
+  ethereum: Ethereum;
+};